603: Renamed otc store to otc-make store
